export_collection_orchestrations/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coverage==7.2.1
supabase~=1.0.0
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pymongo==4.3.3
pytz==2022.7.1
requests==2.28.2
sqlparse==0.2.4
tzdata==2022.7
urllib3==1.26.15